ex05 updated

+++ ex05/ft_putstr.c
@@ -6,12 +6,11 @@
 /*   By: ukhoba <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/06/22 09:30:49 by ukhoba            #+#    #+#             */
-/*   Updated: 2020/06/22 09:30:55 by ukhoba           ###   ########.fr       */
+/*   Updated: 2020/06/23 08:37:42 by ukhoba           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <unistd.h>
-#include <string.h>
 
 void	ft_putchar(char c)
 {
@@ -20,7 +19,7 @@ void	ft_putchar(char c)
 
 void	ft_putstr(char *str)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
